# zip-repl-1/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** Optional machine-readable error code, for programmatic error handling. */
  code?: string;
}

/**
 * Multipart form fields for a Stage 0 run. Fully automatic — the only input is the map image itself.
 */
export interface Stage0RunInput {
  /** Any supported map image (AI-generated, satellite/Google-Maps-style, CAD, colored campus layout, etc.). */
  map: Blob;
}

/**
 * Diagnostics about what Stage 0 detected, at the canonical processing resolution.
 */
export interface Stage0Metadata {
  widthPx: number;
  heightPx: number;
  /** Fraction of the canonical map classified as building footprint. */
  buildingCoverageFraction: number;
  /** Fraction of the canonical map classified as road/walkway (cosmetic only — not used by the Stage 1 physics mask). */
  roadCoverageFraction: number;
  /** Fraction of the canonical map classified as open ground. */
  openAreaFraction: number;
}

export interface Stage0RunSummary {
  runId: string;
  createdAt: string;
  executionTimeMs: number;
  metadata: Stage0Metadata;
}

export interface Stage0Run {
  runId: string;
  createdAt: string;
  executionTimeMs: number;
  /** Standardized, clean engineering-style rendering of the map (Processed_Map.png), as a base64 PNG data URL. */
  processedMapImageBase64: string;
  /** Human-viewable rendering of the detected building/obstruction mask (BinaryMask.png), as a base64 PNG data URL. Stage 1 loads the lossless array form of this mask directly from disk, not this image. */
  binaryMaskImageBase64: string;
  metadata: Stage0Metadata;
}

/**
 * Multipart form fields for a Stage 1 run. Exactly one of `stage0RunId` or `map` must be provided.
 */
export interface Stage1RunInput {
  /** Preferred map source — a completed Stage 0 run to consume directly (no re-reading of the original upload). */
  stage0RunId?: string;
  /** Workflow A — the user has explicitly indicated this map is already in the simulator's native "Simulation Ready" format, so it is used as-is with no preprocessing. Use `stage0RunId` instead (Workflow B) for general map styles. */
  map?: Blob;
  /** Physical size of one grid cell in meters (h). Ignored if realWidthMeters is set. */
  cellSizeMeters?: number;
  /** Optional true physical width of the map in meters, used to auto-calibrate cell size for this map. */
  realWidthMeters?: number;
  /** Simulation frequency in MHz (compute/RAM-saving proxy frequency, not literal 2.4GHz Wi-Fi). */
  frequencyMHz?: number;
  /** Effective refractive index of building material (n_eff). */
  refractiveIndex?: number;
  /** Absorption coefficient of building material in 1/m (alpha_eff). */
  absorptionCoeff?: number;
  /** Transmit amplitude injected at the antenna location. */
  sourceValue?: number;
  /** Bottom of the colorbar scale (display only, does not affect the computed field). */
  minDb?: number;
  /** Width, in grid cells, of the absorbing boundary ramp around the domain edges. */
  pmlWidth?: number;
  /** Strength of the absorbing boundary at the domain edge. */
  pmlMaxLoss?: number;
  /** Optional flat extra loss approximating 3D free-space spreading (0 = off, matches original 2D behavior). */
  alpha3dBump?: number;
  /** Horizontal antenna position as percent across the map (0=left, 100=right). */
  sourceXPercent?: number;
  /** Vertical antenna position as percent down the map (0=top, 100=bottom). */
  sourceYPercent?: number;
  /** Target total grid cell count (rows x cols). Rows/cols are derived from this and the map's aspect ratio. Higher = finer detail, more compute time. */
  cellBudget?: number;
}

/**
 * The fully resolved parameters actually used for the run (after defaults and auto-calibration).
 */
export interface Stage1ResolvedParams {
  cellSizeMeters: number;
  frequencyMHz: number;
  refractiveIndex: number;
  absorptionCoeff: number;
  sourceValue: number;
  minDb: number;
  pmlWidth: number;
  pmlMaxLoss: number;
  alpha3dBump: number;
  sourceXPercent: number;
  sourceYPercent: number;
  cellBudget: number;
}

export interface Stage1RunSummary {
  runId: string;
  createdAt: string;
  gridRows: number;
  gridCols: number;
  peakDb: number;
  executionTimeMs: number;
}

/**
 * Whether this run consumed a Stage 0 processed map ("stage0") or a directly-uploaded compatible raw map ("raw").
 */
export type Stage1RunSourceMode = typeof Stage1RunSourceMode[keyof typeof Stage1RunSourceMode];


export const Stage1RunSourceMode = {
  stage0: 'stage0',
  raw: 'raw',
} as const;

export interface Stage1Run {
  runId: string;
  createdAt: string;
  /** Wall-clock execution time of the simulation, in milliseconds. */
  executionTimeMs: number;
  gridRows: number;
  gridCols: number;
  /** Peak absolute signal strength found in the field, in dB. */
  peakDb: number;
  /** Antenna column index in grid space, after snapping to open space if needed. */
  sourceX: number;
  /** Antenna row index in grid space, after snapping to open space if needed. */
  sourceY: number;
  /** Whether the requested antenna position was inside a building and snapped to the nearest open cell. */
  wasSourceSnapped?: boolean;
  /** Fraction of the grid classified as building/obstacle. */
  occupiedFraction?: number;
  /** Rendered heatmap overlay as a base64 PNG data URL. */
  heatmapImageBase64: string;
  params: Stage1ResolvedParams;
  /** Whether this run consumed a Stage 0 processed map ("stage0") or a directly-uploaded compatible raw map ("raw"). */
  sourceMode: Stage1RunSourceMode;
  /** The Stage 0 run this simulation was based on, if sourceMode is "stage0". */
  stage0RunId?: string;
}

/**
 * Axis-aligned rectangle in heatmap-image pixel coordinates (natural image size, before any CSS scaling) that defines the valid computational domain for Stages 2 and 3.  Everything outside this region is treated as out-of-bounds: no dead zones are detected there, no candidate nodes are placed there.
 */
export interface HeatmapRoi {
  /** Left edge in heatmap image pixels (0 = image left). */
  x: number;
  /** Top edge in heatmap image pixels (0 = image top). */
  y: number;
  /** Rectangle width in heatmap image pixels. */
  width: number;
  /** Rectangle height in heatmap image pixels. */
  height: number;
}

export interface Stage2RunInput {
  /** The Stage 1 run to build on. Stage 1 is not recomputed. */
  stage1RunId: string;
  /** Optional heatmap ROI. When provided, node placement is restricted to this region (colorbar, margins, and titles are excluded automatically). Stored in the run record so Stage 3 can reuse it without asking the user again. */
  roi?: HeatmapRoi;
}

export interface Stage2Run {
  runId: string;
  createdAt: string;
  stage1RunId: string;
  executionTimeMs: number;
  /**
     * @items.minItems 2
     * @items.maxItems 2
     */
  finalNodes: number[][];
  /**
     * @items.minItems 2
     * @items.maxItems 2
     */
  candidateLocations: number[][];
  /**
     * @items.minItems 2
     * @items.maxItems 2
     */
  clusterCentroids: number[][];
  coveragePercent: number;
  nodeCount: number;
  numClusters: number;
  bestExclusionRadius: number;
  nodeCoverageRadius: number;
  /**
     * @minItems 2
     * @maxItems 2
     */
  transmitter: number[];
  imageWidth: number;
  imageHeight: number;
  roi?: HeatmapRoi;
  /** Stage 1 heatmap with optimized node overlay, as a base64 PNG data URL. */
  visualizationImageBase64: string;
}

export interface Stage3RunInput {
  /** The Stage 2 run to build on. Stage 1 and 2 are not recomputed. */
  stage2RunId: string;
  /**
     * Maximum number of nodes the budget-constrained selection may place.
     * @minimum 1
     */
  maxNodes: number;
}