# zip-repl-1/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stage3/runs:
    post:
      operationId: createStage3Run
      tags: [stage3]
      summary: Run Stage 3 — budget-constrained node selection (not yet implemented)
      description: >
        Will consume a Stage 2 run's optimized node set directly (no
        recomputation) and select the best subset under a maximum node budget.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Stage3RunInput"
      responses:
        "501":
          description: Stage 3 is not implemented yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: Optional machine-readable error code, for programmatic error handling.
      required:
        - error
...
    Stage3RunInput:
      type: object
      required:
        - stage2RunId
        - maxNodes
      properties:
        stage2RunId:
          type: string
          description: The Stage 2 run to build on. Stage 1 and 2 are not recomputed.
        maxNodes:
          type: integer
          minimum: 1
          description: Maximum number of nodes the budget-constrained selection may place.